misc,tests: Increase Weekly GPU test timeout
The Weekly GPU tests are failing due to a timeout but I found the testing
timeout was set to 5 hours and  we have been frequently close to reaching this
but have recently changes the test enought o consistently go over.

 The main two things that appear to have caused this are:

1. Moving the X86_VEGA compilation into the the same step as the running of
   the tests.
2. Reducing the number of threads per GitHub Actions runner, thus slowing
   job execution.

In addition we've added more tests to this weekly GPU suite though I don't
believe have got to running these tests yet. The timeout appears to
always been triggered before this.

This PR increases the timout to 3 days and moves the compilation into a
seperate step.