cpu: Add EpisodeCount debug flag (#1861)
Updating the print statement to only print when the debug flag is
specified.

Adding this change as the current gpu0random test is failing due to
being killed by the runner. The test runs as expected when tested
locally.

The test migh be getting killed due to having a very large output.