arch-vega: Fix LDS/buffer load/store x2,x3,x4 (#1916)
These instructions have the wrong index values when writing to the
GPUDynInst d_data field, which expects data to be contiguous in the
d_data buffer. This results in loads/stores being mis-aligned with zeros
padding the empty space.

Set the correct index values (see flat.cc for example) so that the data
values are in the correct locations. This should help fix issues with
nanoGPT and MNIST spotted during the ISCA tutorial.