sim,mem,cpu,arch-arm: Avoid virtual calls on partially constructed probes
Calling manager->addListener() in the ProbeListener constructor means
we are passing a partially constructed object to the addListener()
function. I noticed this because I tried to add a dynamic_cast<> inside
addListener() to assert the function signatures match and it was not
working as expected since the object passed in has a vtable pointer
that points to the base class rather than the derived class during the
partially constructed state.

This introduces a new ProbeListenerPtr unique_ptr specialization with
a custom deleter that takes care of calling removeListener() when the
listener is destroyed. It also adds a new connect() method to
ProbeManager to ensure that the addListener() method is only called
once the ProbeListener is fully constructed.

Change-Id: Ic9b2f22c148e8bd66fd3d04c7a32ac6cb2b041a5