mem-ruby: Move GPU L1 cache MSHR to the coalescer
This commit undoes PR #540 which added MSHRs to the L1 SQC/TCP to handle
requests to the same line. The commit moves the MSHR instead to the
coalescer where the coalescedTable maintains a list of all requests that
go to the same line. These pending requests are completed as soon as the
first request that went to cache returns. If any of these pending
requests have GLC/SLC flags set, then the request is not marked
complete. It is instead sent to cache.