gpu-compute: Implement kernarg preload
Kernarg preloading is a way to pre-populate a wavefront's user SGPRs
when the wave is created from device memory, avoiding some s_load_*
instructions:

https://llvm.org/docs/AMDGPUUsage.html#preloaded-kernel-arguments

Implement this feature as it is used in newer versions of ROCm and
is assumed by the compiler to be supported by gfx942.

Tested using hipblaslt-bench which uses the feature. Should fix any
misbehaving application which printed "warn: Kernarg preload not
implemented."