python: assert if max processes not set for multisim (#2090)
This commit adds an assert for when multisim.set_num_processes() is not
called in a config script. Previously, when the maximum number of
processes was not set, the default behavior would be to use all of the
threads available on the host system. A later change to multisim changed
the default behavior, making it so that no simulations would be launched
if the maximum wasn't set. This commit prevents this by making the user
set a maximum number of processes.