arch-riscv: Fix misprediction of control flow instruction caused by vset{i}vl{i} (#1709)
This PR fixed the bug mentioned in #1708.

In the new implementation, the BPU only predicts vl&vtype for
vset{i}vl{i} instructions. Other control flow instructions will not be
predicted by the BPU for vl and vtype, because it is meaningless. The
new implementation adds `new_conf` flag in pcstate and `vl/vtype`
registers in decoder. This is used to filter the prediction of
`vl&vtype` by other control flow instructions. When a flush occurs, the
`vl/vtype` registers in the decoder are restored.

When checking pcstate at IEW stage, `vl&vtype` are only checked for
instructions with new_vconf = 1. It means that vl and vtype are only
checked for vset{i}vl{i} instructions.

Here is a simple picture describing the new changes：

![绘图1](https://github.com/user-attachments/assets/210e1d34-00d2-4111-a389-e1306d8f25e1)

The following picture is the new pipeline of the test case from #1708.
No more flush after warmup.
![截屏2024-10-24 下午4 08
02](https://github.com/user-attachments/assets/9942b2ef-73f2-463b-a982-74950f0e726a)

Change-Id: I7d3df0328d5757f6b3506743200d5b154dee367f

---------

Co-authored-by: Zhibo Hong <hongzhibo@bytedance.com>
Co-authored-by: Harshil Patel <harshilp2107@gmail.com>