misc,tests: Fix weekly tests by removing dummy job
The weekly tests were failing as a job in a GitHub Action cannot be
queued for more than 24 hours. The Matrix Job takes about 1.5 days
to complete meaning the dummy job at the meats the timeout.

This 24 hour queue time-limit for a Action jobs appears to have
been applied to GitHub recently as this wasn't an issue until then.