misc,tests: Increase Daily GPU Tests timeout to 24h
The Daily GPU Tests are failing due to a timeout. At this time it is
unknown whether this is simply due to these tests requiring >12h to
complete or the tests are hanging. While this is being investigated
for a more complete answer this commit increases the timeout to 24h
so the next run of the Daily Tests passes if it is indeed an issue
of the timeout being set too low.