sim: Add warning for non-default create function
It is very difficult to track down the problem when the empty dummy
shunt is chosen. It turns out, this is usually because your SimObject
has pure virtual functions, but the error message from the compiler
simply says the create function can't be found. This commit adds a
warning and explanation for this case.

There is no warning if the developer explicitly says that they are going
to override the create function.

Note: In the future we can remove the confusing template stuff when we
fully deprecate using the default create method in favor of explicitly
marking in the SimObject description file when we are overriding create.