misc,tests: Improve sep of tests with/without docker req
This refactor means any test that requires the gcn-gpu docker
container is more cleanly seperated from the remainder of the test.

`--host gcn_gpu` can be passed to `main.py` to filter tests down to
only those that use the `gcn-gpu` Docker container.
`--exclude-tags gcn-gpu` can be used to exclude tests that require
the container.