cpu-o3: put unsent mem req to retry queue (#2173)
ISSUE: Currently if a load instruction is performed, it is broken into
message requests to be sent to available cache ports which is a
limitation on how many requests can be sent in a single cycle.
If the requests are more than the available ports, then the instruction
is put into the blockedMemInst queue. This makes the instruction to be
re-issued in the next cycle and executed in the cycle after (basically
using an extra cycle to re-issue).

I feel there is a slight misunderstanding in the two queues, namely
blockedMemInsts and retryMemInsts. In the current scenario, we do not
need to put the instruction in the blocked queue. Also, at certain
places the usage of blockedMemInst is not justified. I can clarify more
with the exact lines of code if required.

THIS COMMIT: If a load memory instruction is not sent due to occupied
ports then put it in retryMemInst queue (if cache is unblocked) instead
of blockedMemInst queue. This saves an unnecessary cycle used to
re-issue. The commit adds the necessary declarations and definitions
needed to perform this.

---------

Co-authored-by: Jason Lowe-Power <jason@lowepower.com>