tests: Reduce Nightly/Weekly X86/ARM Boot tests (#2365)
Attemps to reduce the execution time of the X86/ARM boot tests for the
Nightly (long) and Weekly (very-long) suites. The changes reduce the
number and length of tests with the goal of achieving the same coverage.

The main mains to do this are:

- Where appropriate, exclude systemd from the boot, just boot to the
`init` script then exit.
- Remove redundant tests (e.g.., no point in testing 2 cores sim of also
testing 8 core sim with all else equal).